@@ -6,8 +6,8 @@ author: Dart Team <misc@dartlang.org>
 homepage: https://www.github.com/dart-lang/utf
 
 environment:
-  sdk: '>=2.0.0-dev.20.0 <3.0.0'
+  sdk: '>=2.13.1 <3.0.0'
 
 dev_dependencies:
-  test: ^1.2.0
-  pedantic: ^1.8.0
+  pedantic: ^1.11.1
+  test: ^1.17.9
