@@ -34,7 +34,7 @@ const int _UTF8_SUBSEQUENT_BYTE_BASE = 0x80;
 /// throw an ArgumentError rather than replace the bad value.
 IterableUtf8Decoder decodeUtf8AsIterable(List<int> bytes,
     [int offset = 0,
-    int length,
+    int length = 0,
     int replacementCodepoint = UNICODE_REPLACEMENT_CHARACTER_CODEPOINT]) {
   return IterableUtf8Decoder(bytes, offset, length, replacementCodepoint);
 }
@@ -46,19 +46,20 @@ IterableUtf8Decoder decodeUtf8AsIterable(List<int> bytes,
 /// rather than replace the bad value.
 String decodeUtf8(List<int> bytes,
     [int offset = 0,
-    int length,
+    int? length,
     int replacementCodepoint = UNICODE_REPLACEMENT_CHARACTER_CODEPOINT]) {
   return String.fromCharCodes(
-      (Utf8Decoder(bytes, offset, length, replacementCodepoint)).decodeRest());
+      (Utf8Decoder(bytes, offset, length, replacementCodepoint)).decodeRest()
+          as Iterable<int>);
 }
 
 /// Produce a sequence of UTF-8 encoded bytes from the provided string.
-List<int> encodeUtf8(String str) => codepointsToUtf8(stringToCodepoints(str));
+List<int?> encodeUtf8(String str) => codepointsToUtf8(stringToCodepoints(str));
 
-int _addToEncoding(int offset, int bytes, int value, List<int> buffer) {
+int? _addToEncoding(int offset, int bytes, int? value, List<int?> buffer) {
   while (bytes > 0) {
     buffer[offset + bytes] =
-        _UTF8_SUBSEQUENT_BYTE_BASE | (value & _UTF8_LO_SIX_BIT_MASK);
+        _UTF8_SUBSEQUENT_BYTE_BASE | (value! & _UTF8_LO_SIX_BIT_MASK);
     value = value >> 6;
     bytes--;
   }
@@ -66,12 +67,13 @@ int _addToEncoding(int offset, int bytes, int value, List<int> buffer) {
 }
 
 /// Encode code points as UTF-8 code units.
-List<int> codepointsToUtf8(List<int> codepoints, [int offset = 0, int length]) {
+List<int?> codepointsToUtf8(List<int?> codepoints,
+    [int offset = 0, int? length]) {
   var source = ListRange(codepoints, offset, length);
 
   var encodedLength = 0;
   for (var value in source) {
-    if (value < 0 || value > UNICODE_VALID_RANGE_MAX) {
+    if (value! < 0 || value > UNICODE_VALID_RANGE_MAX) {
       encodedLength += 3;
     } else if (value <= _UTF8_ONE_BYTE_MAX) {
       encodedLength++;
@@ -84,10 +86,10 @@ List<int> codepointsToUtf8(List<int> codepoints, [int offset = 0, int length]) {
     }
   }
 
-  var encoded = List<int>(encodedLength);
+  var encoded = List<int?>.filled(encodedLength, 0, growable: true);
   var insertAt = 0;
   for (var value in source) {
-    if (value < 0 || value > UNICODE_VALID_RANGE_MAX) {
+    if (value! < 0 || value > UNICODE_VALID_RANGE_MAX) {
       encoded.setRange(insertAt, insertAt + 3, [0xef, 0xbf, 0xbd]);
       insertAt += 3;
     } else if (value <= _UTF8_ONE_BYTE_MAX) {
@@ -96,17 +98,17 @@ List<int> codepointsToUtf8(List<int> codepoints, [int offset = 0, int length]) {
     } else if (value <= _UTF8_TWO_BYTE_MAX) {
       encoded[insertAt] = _UTF8_FIRST_BYTE_OF_TWO_BASE |
           (_UTF8_FIRST_BYTE_OF_TWO_MASK &
-              _addToEncoding(insertAt, 1, value, encoded));
+              _addToEncoding(insertAt, 1, value, encoded)!);
       insertAt += 2;
     } else if (value <= _UTF8_THREE_BYTE_MAX) {
       encoded[insertAt] = _UTF8_FIRST_BYTE_OF_THREE_BASE |
           (_UTF8_FIRST_BYTE_OF_THREE_MASK &
-              _addToEncoding(insertAt, 2, value, encoded));
+              _addToEncoding(insertAt, 2, value, encoded)!);
       insertAt += 3;
     } else if (value <= UNICODE_VALID_RANGE_MAX) {
       encoded[insertAt] = _UTF8_FIRST_BYTE_OF_FOUR_BASE |
           (_UTF8_FIRST_BYTE_OF_FOUR_MASK &
-              _addToEncoding(insertAt, 3, value, encoded));
+              _addToEncoding(insertAt, 3, value, encoded)!);
       insertAt += 4;
     }
   }
@@ -115,9 +117,9 @@ List<int> codepointsToUtf8(List<int> codepoints, [int offset = 0, int length]) {
 
 // Because UTF-8 specifies byte order, we do not have to follow the pattern
 // used by UTF-16 & UTF-32 regarding byte order.
-List<int> utf8ToCodepoints(List<int> utf8EncodedBytes,
+List<int?> utf8ToCodepoints(List<int> utf8EncodedBytes,
     [int offset = 0,
-    int length,
+    int? length,
     int replacementCodepoint = UNICODE_REPLACEMENT_CHARACTER_CODEPOINT]) {
   return Utf8Decoder(utf8EncodedBytes, offset, length, replacementCodepoint)
       .decodeRest();
@@ -128,7 +130,7 @@ List<int> utf8ToCodepoints(List<int> utf8EncodedBytes,
 /// as requested by the user of the iterator. (Note: results are not cached.)
 // TODO(floitsch): Consider removing the extend and switch to implements since
 // that's cheaper to allocate.
-class IterableUtf8Decoder extends IterableBase<int> {
+class IterableUtf8Decoder extends IterableBase<int?> {
   final List<int> bytes;
   final int offset;
   @override
@@ -137,7 +139,7 @@ class IterableUtf8Decoder extends IterableBase<int> {
 
   IterableUtf8Decoder(this.bytes,
       [this.offset = 0,
-      this.length,
+      this.length = 0,
       this.replacementCodepoint = UNICODE_REPLACEMENT_CHARACTER_CODEPOINT]);
 
   @override
@@ -151,23 +153,24 @@ class IterableUtf8Decoder extends IterableBase<int> {
 /// character. Set the replacementCharacter to null to throw an
 /// ArgumentError rather than replace the bad value. The return value
 /// from this method can be used as an Iterable (e.g. in a for-loop).
-class Utf8Decoder implements Iterator<int> {
+class Utf8Decoder implements Iterator<int?> {
   // TODO(kevmoo): should this field be private?
   final ListRangeIterator utf8EncodedBytesIterator;
   final int replacementCodepoint;
-  int _current;
+  int? _current;
 
   Utf8Decoder(List<int> utf8EncodedBytes,
       [int offset = 0,
-      int length,
+      int? length,
       this.replacementCodepoint = UNICODE_REPLACEMENT_CHARACTER_CODEPOINT])
       : utf8EncodedBytesIterator =
             (ListRange(utf8EncodedBytes, offset, length)).iterator;
 
   // Decode the remainder of the characters in this decoder
   //into a [List<int>].
-  List<int> decodeRest() {
-    var codepoints = List<int>(utf8EncodedBytesIterator.remaining);
+  List<int?> decodeRest() {
+    var codepoints = List<int?>.filled(utf8EncodedBytesIterator.remaining, 0,
+        growable: true);
     var i = 0;
     while (moveNext()) {
       codepoints[i++] = current;
@@ -175,14 +178,14 @@ class Utf8Decoder implements Iterator<int> {
     if (i == codepoints.length) {
       return codepoints;
     } else {
-      var truncCodepoints = List<int>(i);
+      var truncCodepoints = List<int?>.filled(i, 0, growable: true);
       truncCodepoints.setRange(0, i, codepoints);
       return truncCodepoints;
     }
   }
 
   @override
-  int get current => _current;
+  int? get current => _current;
 
   @override
   bool moveNext() {
@@ -190,7 +193,7 @@ class Utf8Decoder implements Iterator<int> {
 
     if (!utf8EncodedBytesIterator.moveNext()) return false;
 
-    var value = utf8EncodedBytesIterator.current;
+    var value = utf8EncodedBytesIterator.current!;
     var additionalBytes = 0;
 
     if (value < 0) {
@@ -236,7 +239,7 @@ class Utf8Decoder implements Iterator<int> {
     }
     var j = 0;
     while (j < additionalBytes && utf8EncodedBytesIterator.moveNext()) {
-      var nextValue = utf8EncodedBytesIterator.current;
+      var nextValue = utf8EncodedBytesIterator.current!;
       if (nextValue > _UTF8_ONE_BYTE_MAX &&
           nextValue < _UTF8_FIRST_BYTE_OF_TWO_BASE) {
         value = ((value << 6) | (nextValue & _UTF8_LO_SIX_BIT_MASK));
