@@ -9,14 +9,14 @@ import 'list_range.dart';
 import 'utf_16_code_unit_decoder.dart';
 
 /// Decodes the utf16 codeunits to codepoints.
-List<int> utf16CodeUnitsToCodepoints(List<int> utf16CodeUnits,
+List<int?> utf16CodeUnitsToCodepoints(List<int?> utf16CodeUnits,
     [int offset = 0,
-    int length,
+    int? length,
     int replacementCodepoint = UNICODE_REPLACEMENT_CHARACTER_CODEPOINT]) {
   var source = (ListRange(utf16CodeUnits, offset, length)).iterator;
   var decoder =
       Utf16CodeUnitDecoder.fromListRangeIterator(source, replacementCodepoint);
-  var codepoints = List<int>(source.remaining);
+  var codepoints = List<int?>.filled(source.remaining, 0, growable: true);
   var i = 0;
   while (decoder.moveNext()) {
     codepoints[i++] = decoder.current;
@@ -24,21 +24,21 @@ List<int> utf16CodeUnitsToCodepoints(List<int> utf16CodeUnits,
   if (i == codepoints.length) {
     return codepoints;
   } else {
-    var codepointTrunc = List<int>(i);
+    var codepointTrunc = List<int?>.filled(i, 0, growable: true);
     codepointTrunc.setRange(0, i, codepoints);
     return codepointTrunc;
   }
 }
 
 /// Encode code points as UTF16 code units.
-List<int> codepointsToUtf16CodeUnits(List<int> codepoints,
+List<int?> codepointsToUtf16CodeUnits(List<int> codepoints,
     [int offset = 0,
-    int length,
+    int? length,
     int replacementCodepoint = UNICODE_REPLACEMENT_CHARACTER_CODEPOINT]) {
   var listRange = ListRange(codepoints, offset, length);
   var encodedLength = 0;
   for (var value in listRange) {
-    if ((value >= 0 && value < UNICODE_UTF16_RESERVED_LO) ||
+    if ((value! >= 0 && value < UNICODE_UTF16_RESERVED_LO) ||
         (value > UNICODE_UTF16_RESERVED_HI && value <= UNICODE_PLANE_ONE_MAX)) {
       encodedLength++;
     } else if (value > UNICODE_PLANE_ONE_MAX &&
@@ -49,10 +49,10 @@ List<int> codepointsToUtf16CodeUnits(List<int> codepoints,
     }
   }
 
-  var codeUnitsBuffer = List<int>(encodedLength);
+  var codeUnitsBuffer = List<int?>.filled(encodedLength, 0, growable: true);
   var j = 0;
   for (var value in listRange) {
-    if ((value >= 0 && value < UNICODE_UTF16_RESERVED_LO) ||
+    if ((value! >= 0 && value < UNICODE_UTF16_RESERVED_LO) ||
         (value > UNICODE_UTF16_RESERVED_HI && value <= UNICODE_PLANE_ONE_MAX)) {
       codeUnitsBuffer[j++] = value;
     } else if (value > UNICODE_PLANE_ONE_MAX &&
